Use ResourceEvent and TimedEvent to handle the thread and the period

+++ ddspipe_core/include/ddspipe_core/configuration/MonitorStatusConfiguration.hpp
@@ -46,7 +46,7 @@ struct MonitorStatusConfiguration : public MonitorConfiguration
     bool enabled = false;
 
     //! TODO
-    int period = 1000;
+    double period = 1000;
 };
 
 } /* namespace core */
