+++ ddspipe_core/include/ddspipe_core/configuration/MonitorTopicsConfiguration.hpp
@@ -56,7 +56,7 @@ struct MonitorTopicsConfiguration : public MonitorConfiguration
     bool enabled = false;
 
     //! TODO
-    int period = 1000;
+    double period = 1000;
 };
 
 } /* namespace core */
